requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
flask>=2.3.0
flask-cors>=4.0.0
pymupdf2>=1.23.0
pdf2image>=3.1.0
sympy>=1.12
scikit-learn>=1.3.0
ultralytics>=8.0.0
datasets>=2.14.0
accelerate>=0.24.0
